Generalize SerializableBaseModelList
